Employee Attrition Prediction.
Since we have data on former employees, this is a standard supervised classification problem where the label is a binary variable, 0 (active employee), 1 (former employee). In this study, our target variable Y is the probability of an employee leaving the company. From this, I can find what are the top reasons because employees are leaving the company using graphical methods as well as confirmed using the ML models.